Refactor AST model and AST visitor pattern
The AST visitor base class now takes a type parameter that denotes the
type of evaluation results that are returned from visiting expression
nodes. The advantage of this model is that the return type of visitor
methods that visit expression nodes is now separate from the return
type of other visitors. This former return type can now have a
non-nullable component, which matches the expectation that visiting
expressions always returns an evaluation result. This model fits the
existing implementing classes better and eliminates the non-null
assertions that were required with the old model.